enabled searching by project (non-optimal)
currently it'll only show All and teh projects
whose data is in the report. so if you search for a specific
project, then the next load will be only it and all.

Until we have pre-pulled pull requests we can't easily
know the full list without double-processing... or something.

We'll fix it soon enough.